adds create_list of catcha for missing rethoric; update deps
Because there was edits to the target source document after annotations
were made, hxlite could not conciliate the annotation and failed to
display them. Tried to import the list of annotations but they already
existed in the database as the edge context-id. So, had to "re-create"
them with different anno_id.

In requirements.txt ended up removing requests since it is now required
by catchpy and pip was throwing errors.